api/requirements.txt
# Core API dependencies
fastapi==0.119.1
uvicorn==0.38.0
python-multipart==0.0.20

# ML dependencies - Python 3.13 requires numpy>=2.1.0 for ml-dtypes
# ML dependencies
scikit-learn>=1.3.0
joblib>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

pydantic==2.12.3
starlette==0.48.0
google-cloud-storage>=2.0.0

